💄(frontend) advertise live viewers in full screen
There is no need to display the live info bar when we are advertising it since
all informations in the bar are also presented in the advertisement.
Render only the advertising for viewer until the live has started.
