+++ src/frontend/components/StudentLiveWrapper/index.tsx
@@ -5,16 +5,18 @@ import { ConverseInitializer } from 'components/ConverseInitializer';
 import DashboardVideoLiveJitsi from 'components/DashboardVideoLiveJitsi';
 import { LiveVideoLayout } from 'components/LiveVideoLayout';
 import { LiveVideoPanel } from 'components/LiveVideoPanel';
+import { StudentLiveAdvertising } from 'components/StudentLiveAdvertising';
 import { StudentLiveControlBar } from 'components/StudentLiveControlBar';
 import { StudentLiveInfoBar } from 'components/StudentLiveInfoBar';
+import VideoPlayer from 'components/VideoPlayer';
+import { pollForLive } from 'data/sideEffects/pollForLive';
 import {
   LivePanelItem,
   useLivePanelState,
 } from 'data/stores/useLivePanelState';
 import { useLiveStateStarted } from 'data/stores/useLiveStateStarted';
 import { useParticipantWorkflow } from 'data/stores/useParticipantWorkflow';
-import { Video } from 'types/tracks';
-import { StudentLiveViewerWrapper } from './StudentLiveViewerWrapper';
+import { liveState, Video } from 'types/tracks';
 
 const messages = defineMessages({
   defaultLiveTitle: {
@@ -42,8 +44,9 @@ export const StudentLiveWrapper: React.FC<StudentLiveWrapperProps> = ({
       currentItem: state.currentItem,
       setPanelVisibility: state.setPanelVisibility,
     }));
-  const { isStarted } = useLiveStateStarted((state) => ({
+  const { isStarted, setIsLiveStarted } = useLiveStateStarted((state) => ({
     isStarted: state.isStarted,
+    setIsLiveStarted: state.setIsStarted,
   }));
   const isParticipantOnstage = useParticipantWorkflow(
     (state) => state.accepted,
@@ -78,6 +81,36 @@ export const StudentLiveWrapper: React.FC<StudentLiveWrapperProps> = ({
     }
   }, [video.xmpp, isStarted, currentItem, showPanelTrigger]);
 
+  useEffect(() => {
+    let canceled = false;
+    const poll = async () => {
+      if (
+        isStarted ||
+        !video.urls ||
+        !video.live_state ||
+        [liveState.IDLE, liveState.STARTING].includes(video.live_state)
+      ) {
+        return;
+      }
+
+      await pollForLive(video.urls);
+      if (canceled) {
+        return;
+      }
+
+      setIsLiveStarted(true);
+    };
+
+    poll();
+    return () => {
+      canceled = true;
+    };
+  }, [video, isStarted]);
+
+  if (!isStarted && !isParticipantOnstage) {
+    return <StudentLiveAdvertising video={video} />;
+  }
+
   return (
     <ConverseInitializer video={video}>
       <LiveVideoLayout
@@ -95,7 +128,11 @@ export const StudentLiveWrapper: React.FC<StudentLiveWrapperProps> = ({
           isParticipantOnstage ? (
             <DashboardVideoLiveJitsi video={video} />
           ) : (
-            <StudentLiveViewerWrapper video={video} playerType={playerType} />
+            <VideoPlayer
+              playerType={playerType}
+              timedTextTracks={[]}
+              video={video}
+            />
           )
         }
         sideElement={<LiveVideoPanel video={video} />}
